Refactor Xilinx test to reuse infrastructure for Intel
